Create pytorch subdir under each model and move loader, src into it and update README.md
