+++ README.md
@@ -8,23 +8,30 @@ This repository contains model implementations that are used for testing and ben
 
 ## Structure
 
+The repository is organized with models as top-level directories, and frameworks as subdirectories. This structure allows supporting multiple implementations of the same model across different frameworks (PyTorch, TensorFlow, JAX, etc.).
+
 ```
-├── base.py                 # Base interface for all model loaders
-├── yolov4/                 # YOLOv4 model implementation
-│   ├── __init__.py         # Package initialization
-│   ├── loader.py           # Public interface - ModelLoader class
-│   └── src/                # Model implementation details
-│       ├── yolov4.py       # Main model implementation
-│       ├── downsample1.py  # Model components
-│       └── ...
+├── base.py                    # Base interface for all model loaders
+├── yolov4/                    # YOLOv4 model implementation
+│   ├── __init__.py            # Package initialization with imports from pytorch subdirectory
+│   └── pytorch/               # PyTorch-specific implementation
+│       ├── __init__.py        # Package initialization
+│       ├── loader.py          # Public interface - ModelLoader class
+│       └── src/               # Model implementation details
+│           ├── yolov4.py      # Main model implementation
+│           ├── downsample1.py # Model components
+│           └── ...
 ├── yolov3/
 │   ├── __init__.py
-│   ├── loader.py
-│   └── src/
-├── oft/
+│   └── pytorch/
+│       ├── __init__.py
+│       ├── loader.py
+│       └── src/
+├── bert/
 │   ├── __init__.py
-│   ├── loader.py
-│   └── src/
+│   └── pytorch/
+│       ├── __init__.py
+│       └── loader.py
 └── ...
 ```
 
@@ -33,7 +40,8 @@ This repository contains model implementations that are used for testing and ben
 Each model in tt-forge-models follows the same standardized interface pattern. All loaders use the class name `ModelLoader` for consistency, making it easy to write generic code that works with any model.
 
 ```python
-from third_party.tt_forge_models.yolov4 import ModelLoader
+# Import specifically from the PyTorch implementation
+from third_party.tt_forge_models.yolov4.pytorch import ModelLoader
 
 # Load model with default settings - no parameters needed
 model = ModelLoader.load_model()
@@ -49,16 +57,18 @@ outputs = model(inputs)
 
 To add a new model:
 
-1. Create a new directory for your model
-2. Create a `loader.py` file that implements the ForgeModel interface with a class named `ModelLoader`
-3. Create a `src/` directory for your model implementation files (if needed)
-4. Implement parameter-free `load_model()` and `load_inputs()` methods
+1. Create a new top-level directory for your model (e.g., `mynewmodel/`)
+2. Create a framework-specific subdirectory (e.g., `mynewmodel/pytorch/`)
+3. Create a `loader.py` file within the framework directory that implements the ForgeModel interface with a class named `ModelLoader`
+4. Create a `src/` directory for your model implementation files (if needed)
+5. Create `__init__.py` files in both the top-level and framework directories
+6. Implement parameter-free `load_model()` and `load_inputs()` methods
 
 Example for a new model:
 
 ```python
-# mymodel/loader.py
-from tt_forge_models.base import ForgeModel
+# mymodel/pytorch/loader.py
+from ...base import ForgeModel
 from .src.model import MyModel
 
 class ModelLoader(ForgeModel):
@@ -80,6 +90,25 @@ class ModelLoader(ForgeModel):
         return inputs
 ```
 
+```python
+# mymodel/pytorch/__init__.py
+from .loader import ModelLoader
+```
+
+```python
+# mymodel/__init__.py
+# Import from the PyTorch implementation by default
+from .pytorch import ModelLoader
+```
+
+## Future Considerations
+
+The new framework-aware directory structure is designed to support various future enhancements:
+
+1. **Multiple Framework Implementations**: The same model could be implemented in PyTorch, TensorFlow, JAX, or other frameworks while maintaining a consistent interface.
+
+2. **Model Variants**: Different versions of the same architecture (e.g., small, medium, large) could be implemented and selected via configuration.
+
 ## Supporting HuggingFace Models
 
 For models available through the HuggingFace Transformers library, we create wrapper loader classes that delegate to the HF APIs. This approach leverages HuggingFace's infrastructure while providing a standardized interface.
@@ -87,10 +116,10 @@ For models available through the HuggingFace Transformers library, we create wra
 Example of BERT implementation:
 
 ```python
-# bert/loader.py
+# bert/pytorch/loader.py
 import torch
 from transformers import AutoTokenizer, AutoModelForQuestionAnswering
-from tt_forge_models.base import ForgeModel
+from ...base import ForgeModel
 
 class ModelLoader(ForgeModel):
     # Shared configuration parameters
@@ -134,12 +163,12 @@ When this repository is used as a git submodule or dependency in other projects,
 
 ```python
 # All models use the consistent ModelLoader class name
-from third_party.tt_forge_models.yolov4 import ModelLoader
-from third_party.tt_forge_models.bert import ModelLoader
+from third_party.tt_forge_models.yolov4.pytorch import ModelLoader
+from third_party.tt_forge_models.bert.pytorch import ModelLoader
 
 # To use multiple models in the same file, use aliases
-from third_party.tt_forge_models.yolov4 import ModelLoader as YOLOv4Loader
-from third_party.tt_forge_models.bert import ModelLoader as BertLoader
+from third_party.tt_forge_models.yolov4.pytorch import ModelLoader as YOLOv4Loader
+from third_party.tt_forge_models.bert.pytorch import ModelLoader as BertLoader
 
 # For base classes and utilities
 from third_party.tt_forge_models.base import ForgeModel
