+++ bert/__init__.py
@@ -4,4 +4,5 @@
 """
 BERT model implementation for Tenstorrent projects.
 """
-from .loader import ModelLoader
+# Import from the PyTorch implementation by default
+from .pytorch import ModelLoader
