+++ yolov4/pytorch/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+"""
+YOLOv4 PyTorch model implementation for Tenstorrent projects.
+"""
+from .loader import ModelLoader
